Conflicts:
	TP-Pokemon-Solution/.vs/TP-Pokemon-Solution/v14/.suo
	TP-Pokemon-Solution/TP-Pokemon/MainWindow.xaml.cs
	TP-Pokemon-Solution/TP-Pokemon/TP-Pokemon.csproj

	TP-Pokemon-Solution/TP-Pokemon/obj/Debug/TP-Pokemon_MarkupCompile.i.cache

+++ TP-Pokemon-Solution/TP-Pokemon/Item.cs
@@ -13,7 +13,7 @@ namespace TP_Pokemon
         public string desItem;
         public TypeMonster typeItem;
         public string valueItem;
-        public int costItem;
+        public float costItem;
 
     }
 }
